fix(storage): ignore active COSI access updates

+++ drivers/cosi-provisioner-sidecar/pkg/bucketaccess/bucketaccess_controller.go
@@ -327,6 +327,10 @@ func (bal *BucketAccessListener) Update(ctx context.Context, old, new *v1alpha1.
 		"name", old.ObjectMeta.Name)
 
 	bucketAccess := new.DeepCopy()
+	if bucketAccess.GetDeletionTimestamp().IsZero() {
+		return nil
+	}
+
 	err := bal.deleteBucketAccessOp(ctx, bucketAccess)
 	if err != nil {
 		return err
